cython: fix for numpy
The most recent cython changes caused mismatches between INTDTYPE_t and long, so let's explicitly declare that we want int32.

Fixes: https://github.com/internetarchive/archive-pdf-tools/issues/75

+++ cython/optimiser.pyx
@@ -8,7 +8,7 @@ import numpy as np
 cimport numpy as np
 cimport cython
 
-INTDTYPE = np.int_
+INTDTYPE = np.int32
 UINT8DTYPE = np.uint8
 ctypedef np.int32_t INTDTYPE_t
 ctypedef np.uint8_t UINT8DTYPE_t
