+++ cython/sauvola.pyx
@@ -13,7 +13,7 @@ import numpy as np
 cimport numpy as np
 cimport cython
 
-INTDTYPE = np.int_
+INTDTYPE = np.int32
 UINT8DTYPE = np.uint8
 
 ctypedef np.int32_t INTDTYPE_t
